refactor: basic tracer tests
