+++ packages/opentelemetry-tracer-basic/test/Span.test.ts
@@ -16,14 +16,13 @@
 
 import * as assert from 'assert';
 import { performance } from 'perf_hooks';
-import { Span } from '../src/Span';
 import {
   SpanKind,
   CanonicalCode,
   TraceFlags,
   SpanContext,
 } from '@opentelemetry/types';
-import { BasicTracer } from '../src';
+import { BasicTracer, Span } from '../src';
 import {
   hrTimeToNanoseconds,
   hrTimeToMilliseconds,
